Make packageJson updates generic
